epdk-platform/ingest/requirements.txt
pandas==2.2.0
xlrd==2.0.1
openpyxl==3.1.2
psycopg2-binary==2.9.9
